Adding unit tests for Provider Feedback Email Service
