+++ src/Sfa.Tl.Matching.Application.UnitTests/Services/ProviderQuarterlyUpdateEmailService/Builders/BackgroundProcessHistoryBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using Sfa.Tl.Matching.Domain.Models;
 using Sfa.Tl.Matching.Models.Enums;
 
-namespace Sfa.Tl.Matching.Application.UnitTests.Services.ProviderFeedback.Builders
+namespace Sfa.Tl.Matching.Application.UnitTests.Services.ProviderQuarterlyUpdateEmailService.Builders
 {
     public class BackgroundProcessHistoryBuilder
     {
